+++ isol/server-common/src/main/java/shinhan/server_common/domain/entity/TempUserRepository.java
@@ -1,7 +1,6 @@
-package shinhan.server_common.domain.tempuser.repository;
+package shinhan.server_common.domain.entity;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import shinhan.server_common.domain.tempuser.TempUser;
 
 public interface TempUserRepository extends JpaRepository<TempUser, Integer> {
     TempUser findBySerialNumber(Long serialNumber);
