+++ isol/server-common/src/main/java/shinhan/server_common/notification/service/SSEService.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
-import shinhan.server_common.domain.tempuser.TempUser;
-import shinhan.server_common.domain.tempuser.repository.TempUserRepository;
+import shinhan.server_common.domain.entity.TempUser;
+import shinhan.server_common.domain.entity.TempUserRepository;
 import shinhan.server_common.global.exception.CustomException;
 import shinhan.server_common.global.exception.ErrorCode;
 import shinhan.server_common.notification.dto.NotificationFindAllResponse;
